pyproject.toml
[tool.poetry]
name = "tg-chatgpt"
version = "0.1.0"
description = ""
authors = ["trover97"]
readme = "README.md"
#packages = [{include = "tg_chatgpt", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
pytelegrambotapi = "^4.10.0"
requests = "^2.28.2"
flask = "^2.2.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
